+++ geomesa-blobstore/geomesa-blobstore-accumulo/src/test/resources/log4j.xml
