+++ geomesa-blobstore/geomesa-blobstore-accumulo/src/test/resources/testFile3.txt
