+++ geomesa-blobstore/geomesa-blobstore-handlers/geomesa-blobstore-exif-handler/pom.xml
@@ -26,8 +26,7 @@
 
         <dependency>
             <groupId>org.locationtech.geomesa</groupId>
-            <artifactId>geomesa-blobstore-core</artifactId>
-            <version>${project.version}</version>
+            <artifactId>geomesa-blobstore-api</artifactId>
         </dependency>
 
         <dependency>
