+++ geomesa-blobstore/geomesa-blobstore-web/pom.xml
@@ -25,7 +25,12 @@
         </dependency>
         <dependency>
             <groupId>org.locationtech.geomesa</groupId>
-            <artifactId>geomesa-blobstore-core</artifactId>
+            <artifactId>geomesa-blobstore-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.locationtech.geomesa</groupId>
+            <artifactId>geomesa-blobstore-accumulo</artifactId>
+            <version>${project.version}</version>
         </dependency>
         <dependency>
             <groupId>org.locationtech.geomesa</groupId>
